Remove ReplAuth and simplify startup
